Fix bug on video export producing wrong output fps
Issue comes from method CVideoIO::getImage(). Each call to this function
after a call to CVideoIO::startRead() leads to grab new frame from
source. However, in many methods, we call getImage() to get the current
image and not the next frame. So we add the method setFrameToRead() to
set the frame index we want to process and grab from the source only if
necessary.